Move instruction preload helper out of core
